refactor(gpu): split IR and compiler modules
