+++ src/gpu/ir/mod.rs
@@ -1,184 +1,18 @@
-//! Minimal semantic GPU IR used by the initial backend contract.
+//! Minimal semantic GPU IR used by the backend contract.
 //!
 //! This is intentionally not a device ABI. It contains no `wgpu` handles,
 //! raw pointers, shader bindings, or CPU trait objects. Geometry, materials,
 //! and textures will be added in later IR phases.
 
-pub type GpuFloat = f32;
-pub type GpuIndex = u32;
+pub mod geometry;
+pub mod math;
+pub mod render;
+pub mod texture;
 
-macro_rules! typed_id {
-    ($name:ident) => {
-        #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-        pub struct $name(pub GpuIndex);
-    };
-}
-
-typed_id!(TransformId);
-typed_id!(GeometryId);
-typed_id!(PrimitiveId);
-typed_id!(MaterialId);
-typed_id!(SpectrumId);
-typed_id!(LightId);
-typed_id!(FloatTextureId);
-typed_id!(SpectrumTextureId);
-typed_id!(TextureMappingId);
-typed_id!(ImageId);
-typed_id!(InstanceDefinitionId);
-typed_id!(InstanceId);
-typed_id!(MinMaxNodeId);
-typed_id!(SourceId);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuPoint2(pub [GpuFloat; 2]);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuPoint3(pub [GpuFloat; 3]);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuVector2(pub [GpuFloat; 2]);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuVector3(pub [GpuFloat; 3]);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuNormal3(pub [GpuFloat; 3]);
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuMatrix4x4(pub [[GpuFloat; 4]; 4]);
-
-impl GpuMatrix4x4 {
-    pub fn identity() -> Self {
-        Self([
-            [1.0, 0.0, 0.0, 0.0],
-            [0.0, 1.0, 0.0, 0.0],
-            [0.0, 0.0, 1.0, 0.0],
-            [0.0, 0.0, 0.0, 1.0],
-        ])
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuMatrix3x3(pub [[GpuFloat; 3]; 3]);
-
-impl GpuMatrix3x3 {
-    pub fn identity() -> Self {
-        Self([[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]])
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuStaticTransform {
-    pub render_from_object: GpuMatrix4x4,
-    pub object_from_render: GpuMatrix4x4,
-    pub swaps_handedness: bool,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuTransform {
-    Static(GpuStaticTransform),
-    Animated(GpuAnimatedTransform),
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuAnimatedTransform {
-    pub start: GpuStaticTransform,
-    pub end: GpuStaticTransform,
-    pub start_time: GpuFloat,
-    pub end_time: GpuFloat,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuBounds3 {
-    pub min: GpuPoint3,
-    pub max: GpuPoint3,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuBounds2 {
-    pub min: GpuPoint2,
-    pub max: GpuPoint2,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct GpuRange {
-    pub offset: GpuIndex,
-    pub count: GpuIndex,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum GpuSpectrumResource {
-    Constant {
-        value: GpuFloat,
-    },
-    PiecewiseLinear {
-        wavelengths_nm: Box<[GpuFloat]>,
-        values: Box<[GpuFloat]>,
-    },
-    RgbAlbedo {
-        coefficients: [GpuFloat; 3],
-    },
-    RgbUnbounded {
-        coefficients: [GpuFloat; 3],
-    },
-    RgbIlluminant {
-        coefficients: [GpuFloat; 3],
-        illuminant: SpectrumId,
-    },
-    Blackbody {
-        temperature_kelvin: GpuFloat,
-        scale: GpuFloat,
-    },
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuImageChannels {
-    R,
-    Rg,
-    Rgb,
-    Rgba,
-}
-
-impl GpuImageChannels {
-    pub fn count(self) -> usize {
-        match self {
-            Self::R => 1,
-            Self::Rg => 2,
-            Self::Rgb => 3,
-            Self::Rgba => 4,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum GpuTexelStorage {
-    U8(Box<[u8]>),
-    F16(Box<[u16]>),
-    F32(Box<[GpuFloat]>),
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuColorEncoding {
-    Linear,
-    Srgb,
-    Gamma { exponent: GpuFloat },
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct GpuMipLevel {
-    pub resolution: [GpuIndex; 2],
-    pub texel_offset: u64,
-    pub texel_count: u64,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuImageResource {
-    pub resolution: [GpuIndex; 2],
-    pub channels: GpuImageChannels,
-    pub storage: GpuTexelStorage,
-    pub mip_levels: Box<[GpuMipLevel]>,
-    pub color_encoding: GpuColorEncoding,
-}
+pub use geometry::*;
+pub use math::*;
+pub use render::*;
+pub use texture::*;
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct GpuIrVersion {
@@ -268,248 +102,6 @@ impl GpuBounds2i {
     }
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuRenderRequestError {
-    ZeroSampleCount,
-    SampleRangeOverflow,
-    SampleRangeExceedsSamplesPerPixel,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct GpuRenderRequest {
-    pub sample_start: u64,
-    pub sample_count: u32,
-}
-
-impl GpuRenderRequest {
-    pub fn new(
-        render: &GpuRenderConfig,
-        sample_start: u64,
-        sample_count: u32,
-    ) -> Result<Self, GpuRenderRequestError> {
-        if sample_count == 0 {
-            return Err(GpuRenderRequestError::ZeroSampleCount);
-        }
-        let sample_end = sample_start
-            .checked_add(u64::from(sample_count))
-            .ok_or(GpuRenderRequestError::SampleRangeOverflow)?;
-        if sample_end > u64::from(render.sampler.samples_per_pixel) {
-            return Err(GpuRenderRequestError::SampleRangeExceedsSamplesPerPixel);
-        }
-        Ok(Self {
-            sample_start,
-            sample_count,
-        })
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuRenderOutput {
-    pub pixel_bounds: GpuBounds2i,
-    pub rgb: Box<[[f32; 3]]>,
-    pub sample_start: u64,
-    pub sample_count: u32,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuRenderOutputError {
-    InvalidPixelBounds,
-    PixelCountMismatch { expected: usize, actual: usize },
-}
-
-impl GpuRenderOutput {
-    pub fn new(
-        pixel_bounds: GpuBounds2i,
-        rgb: Box<[[f32; 3]]>,
-        request: GpuRenderRequest,
-    ) -> Result<Self, GpuRenderOutputError> {
-        let expected = pixel_bounds
-            .pixel_count()
-            .filter(|count| *count > 0)
-            .ok_or(GpuRenderOutputError::InvalidPixelBounds)?;
-        if rgb.len() != expected {
-            return Err(GpuRenderOutputError::PixelCountMismatch {
-                expected,
-                actual: rgb.len(),
-            });
-        }
-        Ok(Self {
-            pixel_bounds,
-            rgb,
-            sample_start: request.sample_start,
-            sample_count: request.sample_count,
-        })
-    }
-}
-
-impl GpuBounds2i {
-    pub fn area(self) -> Option<u64> {
-        let width = u64::from(self.max[0]).checked_sub(u64::from(self.min[0]))?;
-        let height = u64::from(self.max[1]).checked_sub(u64::from(self.min[1]))?;
-        (width > 0 && height > 0).then(|| width.checked_mul(height))?
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuPerspectiveCamera {
-    pub render_from_camera: TransformId,
-    pub camera_from_raster: GpuMatrix4x4,
-    pub lens_radius: GpuFloat,
-    pub focal_distance: GpuFloat,
-    pub shutter_open: GpuFloat,
-    pub shutter_close: GpuFloat,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuIndependentSampler {
-    pub samples_per_pixel: u32,
-    pub seed: u64,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuRgbFilm {
-    pub full_resolution: [GpuIndex; 2],
-    pub pixel_bounds: GpuBounds2i,
-    pub diagonal_mm: GpuFloat,
-    pub output_rgb_from_xyz: GpuMatrix3x3,
-    pub iso: GpuFloat,
-    pub max_component_value: GpuFloat,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuBoxFilter {
-    pub radius: GpuVector2,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuWavefrontVolPath {
-    pub max_depth: u32,
-    pub regularize: bool,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuLightSampler {
-    Uniform,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuRenderConfig {
-    pub camera: GpuPerspectiveCamera,
-    pub sampler: GpuIndependentSampler,
-    pub film: GpuRgbFilm,
-    pub filter: GpuBoxFilter,
-    pub integrator: GpuWavefrontVolPath,
-    pub light_sampler: GpuLightSampler,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuTriangleMesh {
-    pub positions: Vec<GpuPoint3>,
-    pub indices: Vec<[GpuIndex; 3]>,
-    pub normals: Option<Vec<GpuNormal3>>,
-    pub tangents: Option<Vec<GpuVector3>>,
-    pub uvs: Option<Vec<GpuPoint2>>,
-    pub face_indices: Option<Vec<GpuIndex>>,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuBilinearPatchMesh {
-    pub positions: Vec<GpuPoint3>,
-    pub indices: Vec<[GpuIndex; 4]>,
-    pub normals: Option<Vec<GpuNormal3>>,
-    pub uvs: Option<Vec<GpuPoint2>>,
-    pub face_indices: Option<Vec<GpuIndex>>,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuCurveType {
-    Flat,
-    Cylinder,
-    Ribbon,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct GpuCurveSegment {
-    pub control_points: [GpuPoint3; 4],
-    pub widths: [GpuFloat; 2],
-    pub endpoint_normals: Option<[GpuNormal3; 2]>,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuCurveMesh {
-    pub curve_type: GpuCurveType,
-    pub curves: Vec<GpuCurveSegment>,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuQuadric {
-    Sphere {
-        radius: GpuFloat,
-        z_min: GpuFloat,
-        z_max: GpuFloat,
-        phi_max_radians: GpuFloat,
-    },
-    Cylinder {
-        radius: GpuFloat,
-        z_min: GpuFloat,
-        z_max: GpuFloat,
-        phi_max_radians: GpuFloat,
-    },
-    Disk {
-        height: GpuFloat,
-        radius: GpuFloat,
-        inner_radius: GpuFloat,
-        phi_max_radians: GpuFloat,
-    },
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuMinMaxNode {
-    pub parameter_bounds: GpuBounds2,
-    pub displacement_min: GpuFloat,
-    pub displacement_max: GpuFloat,
-    pub children: Option<[MinMaxNodeId; 4]>,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuDisplacedTriangleMesh {
-    pub base_mesh: GeometryId,
-    pub displacement: FloatTextureId,
-    pub displacement_scale: GpuFloat,
-    pub displacement_offset: GpuFloat,
-    pub edge_length: GpuFloat,
-    pub min_max_nodes: Box<[GpuMinMaxNode]>,
-    pub triangle_roots: Box<[MinMaxNodeId]>,
-    pub displaced_bounds_object: Box<[GpuBounds3]>,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum GpuGeometry {
-    TriangleMesh(GpuTriangleMesh),
-    BilinearPatchMesh(GpuBilinearPatchMesh),
-    CurveMesh(GpuCurveMesh),
-    Quadric(GpuQuadric),
-    DisplacedTriangleMesh(GpuDisplacedTriangleMesh),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum GpuAreaLightBinding {
-    None,
-    Uniform(LightId),
-    PerElement(Vec<LightId>),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct GpuPrimitive {
-    pub geometry: GeometryId,
-    pub transform: TransformId,
-    pub material: Option<MaterialId>,
-    pub alpha: Option<FloatTextureId>,
-    pub shadow_alpha: Option<FloatTextureId>,
-    pub area_light: GpuAreaLightBinding,
-    pub reverse_orientation: bool,
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct GpuInstanceDefinition {
     pub primitives: Vec<PrimitiveId>,
@@ -533,96 +125,6 @@ pub struct GpuDiffuseMaterial {
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum GpuMaterial {
     Diffuse(GpuDiffuseMaterial),
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuImageWrapMode {
-    Black,
-    Clamp,
-    Repeat,
-    OctahedralSphere,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuImageFilter {
-    Point,
-    Bilinear,
-    Trilinear,
-    Ewa { max_anisotropy: GpuFloat },
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuTextureMapping {
-    Uv {
-        su: GpuFloat,
-        sv: GpuFloat,
-        du: GpuFloat,
-        dv: GpuFloat,
-    },
-    Spherical {
-        texture_from_render: GpuMatrix4x4,
-    },
-    Cylindrical {
-        texture_from_render: GpuMatrix4x4,
-    },
-    Planar {
-        texture_from_render: GpuMatrix4x4,
-        vs: GpuVector3,
-        vt: GpuVector3,
-        ds: GpuFloat,
-        dt: GpuFloat,
-    },
-    Transform3D {
-        texture_from_render: GpuMatrix4x4,
-    },
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuFloatTexture {
-    Constant {
-        value: GpuFloat,
-    },
-    Image {
-        image: ImageId,
-        mapping: TextureMappingId,
-        scale: GpuFloat,
-        invert: bool,
-        swrap: GpuImageWrapMode,
-        twrap: GpuImageWrapMode,
-        filter: GpuImageFilter,
-        channel: GpuFloatImageChannel,
-    },
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuFloatImageChannel {
-    Channel0,
-    Alpha,
-    RgbAverage,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum GpuSpectrumTexture {
-    Constant {
-        value: SpectrumId,
-    },
-    Image {
-        image: ImageId,
-        mapping: TextureMappingId,
-        scale: GpuFloat,
-        invert: bool,
-        swrap: GpuImageWrapMode,
-        twrap: GpuImageWrapMode,
-        filter: GpuImageFilter,
-        spectrum_type: GpuSpectrumType,
-    },
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum GpuSpectrumType {
-    Albedo,
-    Unbounded,
-    Illuminant,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq)]
